[TASK] Failing test for Null directive
