--- tests/BuilderUnkownDirective/input/index.html
@@ -1,2 +1,0 @@
-Exception: Exception
-Unknown directive "unknown-directive" for line ".. unknown-directive:: some content"
